move token part fetch to common section
