+++ apps/web/src/state/info/queries/tokens/fetchEthPrices.ts
