+++ apps/web/src/views/V3Info/components/DensityChart/CustomToolTip.tsx
@@ -1,8 +1,8 @@
 import { useTranslation } from '@pancakeswap/localization'
 import { AutoColumn, Text } from '@pancakeswap/uikit'
 import useTheme from 'hooks/useTheme'
+import { PoolData } from 'state/info/types'
 import { styled } from 'styled-components'
-import { PoolData } from '../../types'
 import { formatAmount } from '../../utils/numbers'
 import { LightCard } from '../Card'
 import { RowBetween } from '../Row'
