+++ apps/web/src/views/V3Info/components/DensityChart/index.tsx
@@ -3,12 +3,12 @@ import { Flex, Spinner } from '@pancakeswap/uikit'
 import { FeeAmount, Pool, TICK_SPACINGS, TickMath } from '@pancakeswap/v3-sdk'
 import { useCallback, useEffect, useMemo, useState } from 'react'
 import { Bar, BarChart, Cell, LabelList, ResponsiveContainer, Tooltip, XAxis } from 'recharts'
+import { DensityChartEntry, PoolData } from 'state/info/types'
 import { styled } from 'styled-components'
 import { safeGetAddress } from 'utils'
 import { MAX_UINT128 } from '../../constants'
 import { TickProcessed } from '../../data/pool/tickData'
 import { usePoolData, usePoolTickData } from '../../hooks'
-import { DensityChartEntry, PoolData } from '../../types'
 import { CurrentPriceLabel } from './CurrentPriceLabel'
 import CustomToolTip from './CustomToolTip'
