+++ apps/web/src/views/V3Info/utils/data.ts
@@ -1,14 +1,3 @@
-export const get2DayChange = (valueNow: string, value24HoursAgo: string, value48HoursAgo: string): [number, number] => {
-  // get volume info for both 24 hour periods
-  const currentChange = parseFloat(valueNow) - parseFloat(value24HoursAgo)
-  const previousChange = parseFloat(value24HoursAgo) - parseFloat(value48HoursAgo)
-  const adjustedPercentChange = ((currentChange - previousChange) / previousChange) * 100
-  if (Number.isNaN(adjustedPercentChange) || !Number.isFinite(adjustedPercentChange)) {
-    return [currentChange, 0]
-  }
-  return [currentChange, adjustedPercentChange]
-}
-
 /**
  * get standard percent change between two values
  * @param {*} valueNow
