+++ apps/web/src/views/V3Info/utils/index.ts
@@ -1,6 +1,6 @@
 import dayjs from 'dayjs'
+import { ApiTransaction, Transaction, TransactionType } from 'state/info/types'
 import { safeGetAddress } from 'utils'
-import { ApiPoolData, ApiTransaction, PoolData, Transaction, TransactionType } from '../types'
 
 export function shortenAddress(address: string, chars = 4): string {
   const parsed = safeGetAddress(address)
@@ -14,28 +14,6 @@ export function feeTierPercent(fee: number): string {
 }
 
 export const currentTimestamp = () => new Date().getTime()
-
-export function transformPoolData(item: ApiPoolData): PoolData {
-  return {
-    feeTier: item.feeTier,
-    address: item.id,
-    volumeUSD: parseFloat(item.volumeUSD24h),
-    volumeUSDWeek: parseFloat(item.volumeUSD7d),
-    token0: { ...item.token0, address: item.token0.id, derivedETH: 0 },
-    token1: { ...item.token1, address: item.token1.id, derivedETH: 0 },
-    feeUSD: item.totalFeeUSD,
-    liquidity: parseFloat(item.liquidity),
-    sqrtPrice: parseFloat(item.sqrtPrice),
-    tick: item.tick ?? 0,
-    tvlUSD: parseFloat(item.tvlUSD),
-    token0Price: parseFloat(item.token0Price),
-    token1Price: parseFloat(item.token1Price),
-    tvlToken0: parseFloat(item.tvlToken0),
-    tvlToken1: parseFloat(item.tvlToken1),
-    volumeUSDChange: 0,
-    tvlUSDChange: 0,
-  }
-}
 
 export function transformTransaction(transaction: ApiTransaction): Transaction {
   return {
